+++ saleor/graphql/middleware.py
@@ -3,23 +3,32 @@ from typing import Optional
 import opentracing
 import opentracing.tags
 from django.conf import settings
+from django.contrib.auth import authenticate
 from django.contrib.auth.models import AnonymousUser
 from django.utils.functional import SimpleLazyObject
 from graphql import ResolveInfo
-from graphql_jwt.middleware import JSONWebTokenMiddleware
 
 from ..app.models import App
+from ..core.exceptions import ReadOnlyException
 from ..core.tracing import should_trace
 from .views import API_PATH, GraphQLView
 
 
-class JWTMiddleware(JSONWebTokenMiddleware):
+def get_user(request):
+    if not hasattr(request, "_cached_user"):
+        request._cached_user = authenticate(request=request)
+    return request._cached_user
+
+
+class JWTMiddleware:
     def resolve(self, next, root, info, **kwargs):
         request = info.context
 
-        if not hasattr(request, "user"):
-            request.user = AnonymousUser()
-        return super().resolve(next, root, info, **kwargs)
+        def user():
+            return get_user(request) or AnonymousUser()
+
+        request.user = SimpleLazyObject(lambda: user())
+        return next(root, info, **kwargs)
 
 
 class OpentracingGrapheneMiddleware:
@@ -56,6 +65,51 @@ def app_middleware(next, root, info, **kwargs):
                 if auth_prefix.lower() == prefix:
                     request.app = SimpleLazyObject(lambda: get_app(auth_token))
     return next(root, info, **kwargs)
+
+
+class ReadOnlyMiddleware:
+    ALLOWED_MUTATIONS = [
+        "checkoutAddPromoCode",
+        "checkoutBillingAddressUpdate",
+        "checkoutComplete",
+        "checkoutCreate",
+        "checkoutCustomerAttach",
+        "checkoutCustomerDetach",
+        "checkoutEmailUpdate",
+        "checkoutLineDelete",
+        "checkoutLinesAdd",
+        "checkoutLinesUpdate",
+        "checkoutRemovePromoCode",
+        "checkoutPaymentCreate",
+        "checkoutShippingAddressUpdate",
+        "checkoutShippingMethodUpdate",
+        "tokenCreate",
+        "tokenVerify",
+    ]
+
+    @staticmethod
+    def resolve(next_, root, info, **kwargs):
+        operation = info.operation.operation
+        if operation != "mutation":
+            return next_(root, info, **kwargs)
+
+        # Bypass users authenticated with ROOT_EMAIL
+        request = info.context
+        user = getattr(request, "user", None)
+        if user and not user.is_anonymous:
+            user_email = user.email
+            root_email = getattr(settings, "ROOT_EMAIL", None)
+            if root_email and user_email == root_email:
+                return next_(root, info, **kwargs)
+
+        for selection in info.operation.selection_set.selections:
+            selection_name = str(selection.name.value)
+            blocked = selection_name not in ReadOnlyMiddleware.ALLOWED_MUTATIONS
+            if blocked:
+                raise ReadOnlyException(
+                    "Be aware admin pirate! API runs in read-only mode!"
+                )
+        return next_(root, info, **kwargs)
 
 
 def process_view(self, request, view_func, *args):