+++ saleor/graphql/order/mutations/orders.py
@@ -8,6 +8,7 @@ from ....order import events, models
 from ....order.actions import (
     cancel_order,
     clean_mark_order_as_paid,
+    handle_fully_paid_order,
     mark_order_as_paid,
     order_captured,
     order_refunded,
@@ -19,14 +20,14 @@ from ....order.utils import get_valid_shipping_methods_for_order
 from ....payment import CustomPaymentChoices, PaymentError, gateway
 from ...account.types import AddressInput
 from ...core.mutations import BaseMutation
-from ...core.scalars import Decimal
+from ...core.scalars import UUID, Decimal
 from ...core.types.common import OrderError
+from ...core.utils import validate_required_string_field
 from ...meta.deprecated.mutations import ClearMetaBaseMutation, UpdateMetaBaseMutation
 from ...meta.deprecated.types import MetaInput, MetaPath
 from ...order.mutations.draft_orders import DraftOrderUpdate
 from ...order.types import Order, OrderEvent
 from ...shipping.types import ShippingMethod
-from ...utils import get_user_or_app_from_context
 
 
 def clean_order_update_shipping(order, method):
@@ -283,8 +284,9 @@ class OrderAddNote(BaseMutation):
 
     @classmethod
     def clean_input(cls, _info, _instance, data):
-        message = data["input"]["message"].strip()
-        if not message:
+        try:
+            cleaned_input = validate_required_string_field(data["input"], "message")
+        except ValidationError:
             raise ValidationError(
                 {
                     "message": ValidationError(
@@ -292,17 +294,14 @@ class OrderAddNote(BaseMutation):
                     )
                 }
             )
-        data["input"]["message"] = message
-        return data
+        return cleaned_input
 
     @classmethod
     def perform_mutation(cls, _root, info, **data):
         order = cls.get_node_or_error(info, data.get("id"), only_type=Order)
         cleaned_input = cls.clean_input(info, order, data)
         event = events.order_note_added_event(
-            order=order,
-            user=info.context.user,
-            message=cleaned_input["input"]["message"],
+            order=order, user=info.context.user, message=cleaned_input["message"],
         )
         return OrderAddNote(order=order, event=event)
 
@@ -323,8 +322,7 @@ class OrderCancel(BaseMutation):
     def perform_mutation(cls, _root, info, **data):
         order = cls.get_node_or_error(info, data.get("id"), only_type=Order)
         clean_order_cancel(order)
-        requester = get_user_or_app_from_context(info.context)
-        cancel_order(order=order, user=requester)
+        cancel_order(order=order, user=info.context.user)
         return OrderCancel(order=order)
 
 
@@ -393,8 +391,9 @@ class OrderCapture(BaseMutation):
         try_payment_action(
             order, info.context.user, payment, gateway.capture, payment, amount
         )
-
         order_captured(order, info.context.user, amount, payment)
+        if order.is_fully_paid():
+            handle_fully_paid_order(order)
         return OrderCapture(order=order)
 
 
@@ -465,9 +464,7 @@ class OrderUpdateMeta(UpdateMetaBaseMutation):
         public = True
 
     class Arguments:
-        token = graphene.UUID(
-            description="Token of an object to update.", required=True
-        )
+        token = UUID(description="Token of an object to update.", required=True)
         input = MetaInput(
             description="Fields required to update new or stored metadata item.",
             required=True,
@@ -495,7 +492,7 @@ class OrderClearMeta(ClearMetaBaseMutation):
         public = True
 
     class Arguments:
-        token = graphene.UUID(description="Token of an object to clear.", required=True)
+        token = UUID(description="Token of an object to clear.", required=True)
         input = MetaPath(
             description="Fields required to update new or stored metadata item.",
             required=True,