+++ saleor/graphql/order/types.py
@@ -1,10 +1,11 @@
 import graphene
 from django.core.exceptions import ValidationError
 from graphene import relay
-from graphql_jwt.exceptions import PermissionDenied
 
+from ...core.exceptions import PermissionDenied
 from ...core.permissions import AccountPermissions, OrderPermissions
 from ...core.taxes import display_gross_prices
+from ...graphql.utils import get_user_or_app_from_context
 from ...order import OrderStatus, models
 from ...order.models import FulfillmentStatus
 from ...order.utils import get_order_country, get_valid_shipping_methods_for_order
@@ -17,6 +18,7 @@ from ..core.types.common import Image
 from ..core.types.money import Money, TaxedMoney
 from ..decorators import permission_required
 from ..giftcard.types import GiftCard
+from ..invoice.types import Invoice
 from ..meta.deprecated.resolvers import resolve_meta, resolve_private_meta
 from ..meta.types import ObjectWithMetadata
 from ..payment.types import OrderAction, Payment, PaymentChargeStatusEnum
@@ -50,6 +52,9 @@ class OrderEvent(CountableDjangoObjectType):
     quantity = graphene.Int(description="Number of items.")
     composed_id = graphene.String(description="Composed ID of the Fulfillment.")
     order_number = graphene.String(description="User-friendly number of an order.")
+    invoice_number = graphene.String(
+        description="Number of an invoice related to the order."
+    )
     oversold_items = graphene.List(
         graphene.String, description="List of oversold lines names."
     )
@@ -119,6 +124,10 @@ class OrderEvent(CountableDjangoObjectType):
     @staticmethod
     def resolve_order_number(root: models.OrderEvent, _info):
         return root.order_id
+
+    @staticmethod
+    def resolve_invoice_number(root: models.OrderEvent, _info):
+        return root.parameters.get("invoice_number")
 
     @staticmethod
     def resolve_lines(root: models.OrderEvent, _info):
@@ -302,6 +311,9 @@ class Order(CountableDjangoObjectType):
         ShippingMethod,
         required=False,
         description="Shipping methods that can be used with this order.",
+    )
+    invoices = graphene.List(
+        Invoice, required=False, description="List of order invoices."
     )
     number = graphene.String(description="User-friendly number of an order.")
     is_paid = graphene.Boolean(description="Informs if an order is fully paid.")
@@ -492,6 +504,13 @@ class Order(CountableDjangoObjectType):
             else:
                 shipping_method.price = taxed_price.net
         return available
+
+    @staticmethod
+    def resolve_invoices(root: models.Order, info):
+        requester = get_user_or_app_from_context(info.context)
+        if requester == root.user or requester.has_perm(OrderPermissions.MANAGE_ORDERS):
+            return root.invoices.all()
+        raise PermissionDenied()
 
     @staticmethod
     def resolve_is_shipping_required(root: models.Order, _info):