+++ saleor/graphql/payment/mutations.py
@@ -2,12 +2,15 @@ import graphene
 from django.conf import settings
 from django.core.exceptions import ValidationError
 
+from ...checkout import calculations
+from ...checkout.utils import cancel_active_payments
 from ...core.permissions import OrderPermissions
 from ...core.taxes import zero_taxed_money
 from ...core.utils import get_client_ip
+from ...graphql.checkout.utils import clean_billing_address, clean_checkout_shipping
 from ...payment import PaymentError, gateway, models
 from ...payment.error_codes import PaymentErrorCode
-from ...payment.utils import create_payment
+from ...payment.utils import create_payment, is_currency_supported
 from ..account.i18n import I18nMixin
 from ..account.types import AddressInput
 from ..checkout.types import Checkout
@@ -68,21 +71,24 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
     @classmethod
     def calculate_total(cls, info, checkout):
         checkout_total = (
-            info.context.plugins.calculate_checkout_total(
-                checkout, lines=list(checkout), discounts=info.context.discounts
+            calculations.checkout_total(
+                checkout=checkout,
+                lines=list(checkout),
+                discounts=info.context.discounts,
             )
             - checkout.get_total_gift_cards_balance()
         )
+
         return max(checkout_total, zero_taxed_money(checkout_total.currency))
 
     @classmethod
-    def clean_billing_address(cls, billing_address):
-        if billing_address is None:
+    def clean_shipping_method(cls, checkout):
+        if not checkout.shipping_method:
             raise ValidationError(
                 {
-                    "billing_address": ValidationError(
-                        "No billing address associated with this checkout.",
-                        code=PaymentErrorCode.BILLING_ADDRESS_NOT_SET,
+                    "shipping_method": ValidationError(
+                        "Shipping method not set for this checkout.",
+                        code=PaymentErrorCode.SHIPPING_METHOD_NOT_SET,
                     )
                 }
             )
@@ -101,6 +107,18 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
             )
 
     @classmethod
+    def validate_gateway(cls, gateway_id, currency):
+        if not is_currency_supported(currency, gateway_id):
+            raise ValidationError(
+                {
+                    "gateway": ValidationError(
+                        f"The gateway {gateway_id} does not support checkout currency.",
+                        code=PaymentErrorCode.NOT_SUPPORTED_GATEWAY.value,
+                    )
+                }
+            )
+
+    @classmethod
     def perform_mutation(cls, _root, info, checkout_id, **data):
         checkout_id = from_global_id_strict_type(
             checkout_id, only_type=Checkout, field="checkout_id"
@@ -110,17 +128,23 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
         ).get(pk=checkout_id)
 
         data = data["input"]
+        gateway = data["gateway"]
+
+        cls.validate_gateway(gateway, checkout.currency)
 
         checkout_total = cls.calculate_total(info, checkout)
         amount = data.get("amount", checkout_total.gross.amount)
-
-        cls.clean_billing_address(checkout.billing_address)
+        clean_checkout_shipping(
+            checkout, list(checkout), info.context.discounts, PaymentErrorCode
+        )
+        clean_billing_address(checkout, PaymentErrorCode)
         cls.clean_payment_amount(info, checkout_total, amount)
-
         extra_data = {"customer_user_agent": info.context.META.get("HTTP_USER_AGENT")}
 
+        cancel_active_payments(checkout)
+
         payment = create_payment(
-            gateway=data["gateway"],
+            gateway=gateway,
             payment_token=data["token"],
             total=amount,
             currency=settings.DEFAULT_CURRENCY,
@@ -129,7 +153,7 @@ class CheckoutPaymentCreate(BaseMutation, I18nMixin):
             customer_ip_address=get_client_ip(info.context),
             checkout=checkout,
         )
-        return CheckoutPaymentCreate(payment=payment)
+        return CheckoutPaymentCreate(payment=payment, checkout=checkout)
 
 
 class PaymentCapture(BaseMutation):