+++ saleor/graphql/product/resolvers.py
@@ -22,11 +22,22 @@ def resolve_attributes(info, qs=None, in_category=None, in_collection=None, **_k
     return qs.distinct()
 
 
+def resolve_category_by_slug(slug):
+    return models.Category.objects.filter(slug=slug).first()
+
+
 def resolve_categories(info, level=None, **_kwargs):
     qs = models.Category.objects.prefetch_related("children")
     if level is not None:
         qs = qs.filter(level=level)
     return qs.distinct()
+
+
+def resolve_collection_by_slug(info, slug):
+    requestor = get_user_or_app_from_context(info.context)
+    return (
+        models.Collection.objects.visible_to_user(requestor).filter(slug=slug).first()
+    )
 
 
 def resolve_collections(info, **_kwargs):
@@ -36,6 +47,11 @@ def resolve_collections(info, **_kwargs):
 
 def resolve_digital_contents(info):
     return models.DigitalContent.objects.all()
+
+
+def resolve_product_by_slug(info, slug):
+    requestor = get_user_or_app_from_context(info.context)
+    return models.Product.objects.visible_to_user(requestor).filter(slug=slug).first()
 
 
 def resolve_products(info, stock_availability=None, **_kwargs):