+++ saleor/graphql/product/types/products.py
@@ -2,11 +2,13 @@ from dataclasses import asdict
 from typing import List, Union
 
 import graphene
+from django.conf import settings
 from graphene import relay
 from graphene_federation import key
 from graphql.error import GraphQLError
 
 from ....core.permissions import ProductPermissions
+from ....core.weight import convert_weight_to_default_weight_unit
 from ....product import models
 from ....product.templatetags.product_images import (
     get_product_image_thumbnail,
@@ -20,10 +22,8 @@ from ....product.utils.availability import (
 from ....product.utils.costs import get_margin_for_variant, get_product_costs_data
 from ....warehouse.availability import (
     get_available_quantity,
-    get_available_quantity_for_customer,
     get_quantity_allocated,
     is_product_in_stock,
-    is_variant_in_stock,
 )
 from ...account.enums import CountryCodeEnum
 from ...core.connection import CountableDjangoObjectType
@@ -43,6 +43,9 @@ from ...translations.types import (
 )
 from ...utils import get_database_id
 from ...utils.filters import reporting_period_to_date
+from ...warehouse.dataloaders import (
+    AvailableQuantityByProductVariantIdAndCountryCodeLoader,
+)
 from ...warehouse.types import Stock
 from ..dataloaders import (
     CategoryByIdLoader,
@@ -188,16 +191,19 @@ class ProductVariant(CountableDjangoObjectType):
         required=True,
         description="Quantity of a product available for sale.",
         deprecation_reason=(
-            "Use the stock field instead. This field will be removed after 2020-07-31."
+            "Use the quantityAvailable field instead. "
+            "This field will be removed after 2020-07-31."
         ),
     )
-    price_override = graphene.Field(
+    price = graphene.Field(
         Money,
         description=(
-            "Override the base price of a product if necessary. A value of `null` "
-            "indicates that the default product price is used."
+            "Base price of a product variant. "
+            "This field is restricted for admins. "
+            "Use the pricing field to get the public price for customers."
         ),
     )
+
     pricing = graphene.Field(
         VariantPricingInfo,
         description=(
@@ -238,7 +244,6 @@ class ProductVariant(CountableDjangoObjectType):
     digital_content = graphene.Field(
         DigitalContent, description="Digital content for the product variant."
     )
-
     stocks = graphene.Field(
         graphene.List(Stock),
         description="Stocks for the product variant.",
@@ -246,6 +251,19 @@ class ProductVariant(CountableDjangoObjectType):
             CountryCodeEnum,
             description="Two-letter ISO 3166-1 country code.",
             required=False,
+        ),
+    )
+    quantity_available = graphene.Int(
+        required=True,
+        description="Quantity of a product available for sale in one checkout.",
+        country_code=graphene.Argument(
+            CountryCodeEnum,
+            description=(
+                "Two-letter ISO 3166-1 country code. When provided, the exact quantity "
+                "from a warehouse operating in shipping zones that contain this "
+                "country will be returned. Otherwise, it will return the maximum "
+                "quantity from all shipping zones."
+            ),
         ),
     )
 
@@ -258,10 +276,22 @@ class ProductVariant(CountableDjangoObjectType):
         model = models.ProductVariant
 
     @staticmethod
+    @permission_required(ProductPermissions.MANAGE_PRODUCTS)
     def resolve_stocks(root: models.ProductVariant, info, country_code=None):
         if not country_code:
-            return root.stocks.annotate_available_quantity().all()
-        return root.stocks.annotate_available_quantity().for_country(country_code).all()
+            return root.stocks.annotate_available_quantity()
+        return root.stocks.for_country(country_code).annotate_available_quantity()
+
+    @staticmethod
+    def resolve_quantity_available(
+        root: models.ProductVariant, info, country_code=None
+    ):
+        if not root.track_inventory:
+            return settings.MAX_CHECKOUT_LINE_QUANTITY
+
+        return AvailableQuantityByProductVariantIdAndCountryCodeLoader(
+            info.context
+        ).load((root.id, country_code))
 
     @staticmethod
     @permission_required(ProductPermissions.MANAGE_PRODUCTS)
@@ -270,7 +300,12 @@ class ProductVariant(CountableDjangoObjectType):
 
     @staticmethod
     def resolve_stock_quantity(root: models.ProductVariant, info):
-        return get_available_quantity_for_customer(root, info.context.country)
+        if not root.track_inventory:
+            return settings.MAX_CHECKOUT_LINE_QUANTITY
+
+        return AvailableQuantityByProductVariantIdAndCountryCodeLoader(
+            info.context
+        ).load((root.id, info.context.country))
 
     @staticmethod
     def resolve_attributes(root: models.ProductVariant, info):
@@ -287,8 +322,9 @@ class ProductVariant(CountableDjangoObjectType):
         return root.cost_price
 
     @staticmethod
+    @permission_required(ProductPermissions.MANAGE_PRODUCTS)
     def resolve_price(root: models.ProductVariant, *_args):
-        return root.base_price
+        return root.price
 
     @staticmethod
     def resolve_pricing(root: models.ProductVariant, info):
@@ -326,13 +362,17 @@ class ProductVariant(CountableDjangoObjectType):
 
     @staticmethod
     def resolve_is_available(root: models.ProductVariant, info):
-        country = info.context.country
-        return is_variant_in_stock(root, country)
+        if not root.track_inventory:
+            return True
 
-    @staticmethod
-    @permission_required(ProductPermissions.MANAGE_PRODUCTS)
-    def resolve_price_override(root: models.ProductVariant, *_args):
-        return root.price_override
+        def is_variant_in_stock(available_quantity):
+            return available_quantity > 0
+
+        return (
+            AvailableQuantityByProductVariantIdAndCountryCodeLoader(info.context)
+            .load((root.id, info.context.country))
+            .then(is_variant_in_stock)
+        )
 
     @staticmethod
     @permission_required(ProductPermissions.MANAGE_PRODUCTS)
@@ -384,6 +424,10 @@ class ProductVariant(CountableDjangoObjectType):
     def __resolve_reference(root, _info, **_kwargs):
         return graphene.Node.get_node_from_global_id(_info, root.id)
 
+    @staticmethod
+    def resolve_weight(root: models.ProductVariant, _info, **_kwargs):
+        return convert_weight_to_default_weight_unit(root.weight)
+
 
 @key(fields="id")
 class Product(CountableDjangoObjectType):
@@ -407,7 +451,6 @@ class Product(CountableDjangoObjectType):
     is_available = graphene.Boolean(
         description="Whether the product is in stock and visible or not."
     )
-    base_price = graphene.Field(Money, description="The product's default base price.")
     minimal_variant_price = graphene.Field(
         Money, description="The price of the cheapest variant (including discounts)."
     )
@@ -528,28 +571,6 @@ class Product(CountableDjangoObjectType):
         return root.is_visible and in_stock
 
     @staticmethod
-    @permission_required(ProductPermissions.MANAGE_PRODUCTS)
-    def resolve_base_price(root: models.Product, _info):
-        return root.price
-
-    @staticmethod
-    def resolve_price(root: models.Product, info):
-        context = info.context
-
-        def calculate_price(discounts):
-            price_range = root.get_price_range(discounts)
-            price = info.context.plugins.apply_taxes_to_product(
-                root, price_range.start, info.context.country
-            )
-            return price.net
-
-        return (
-            DiscountsByDateTimeLoader(context)
-            .load(info.context.request_time)
-            .then(calculate_price)
-        )
-
-    @staticmethod
     def resolve_attributes(root: models.Product, info):
         return SelectedAttributesByProductIdLoader(info.context).load(root.id)
 
@@ -604,6 +625,10 @@ class Product(CountableDjangoObjectType):
     @staticmethod
     def __resolve_reference(root, _info, **_kwargs):
         return graphene.Node.get_node_from_global_id(_info, root.id)
+
+    @staticmethod
+    def resolve_weight(root: models.Product, _info, **_kwargs):
+        return convert_weight_to_default_weight_unit(root.weight)
 
 
 @key(fields="id")
@@ -685,6 +710,10 @@ class ProductType(CountableDjangoObjectType):
     @staticmethod
     def __resolve_reference(root, _info, **_kwargs):
         return graphene.Node.get_node_from_global_id(_info, root.id)
+
+    @staticmethod
+    def resolve_weight(root: models.ProductType, _info, **_kwargs):
+        return convert_weight_to_default_weight_unit(root.weight)
 
 
 @key(fields="id")