+++ saleor/order/events.py
@@ -55,6 +55,50 @@ def email_sent_event(
     )
 
 
+def invoice_requested_event(*, order: Order, user: Optional[UserType],) -> OrderEvent:
+    return OrderEvent.objects.create(
+        order=order, type=OrderEvents.INVOICE_REQUESTED, user=user
+    )
+
+
+def invoice_generated_event(
+    *, order: Order, user: Optional[UserType], invoice_number: str,
+) -> OrderEvent:
+    return OrderEvent.objects.create(
+        order=order,
+        type=OrderEvents.INVOICE_GENERATED,
+        user=user,
+        parameters={"invoice_number": invoice_number},
+    )
+
+
+def invoice_updated_event(
+    *,
+    order: Order,
+    user: Optional[UserType],
+    invoice_number: str,
+    url: str,
+    status: str
+) -> OrderEvent:
+    return OrderEvent.objects.create(
+        order=order,
+        type=OrderEvents.INVOICE_UPDATED,
+        user=user,
+        parameters={"invoice_number": invoice_number, "url": url, "status": status},
+    )
+
+
+def invoice_sent_event(
+    *, order: Order, user: Optional[UserType], email: str,
+) -> OrderEvent:
+    return OrderEvent.objects.create(
+        order=order,
+        type=OrderEvents.INVOICE_SENT,
+        user=user,
+        parameters={"email": email},
+    )
+
+
 def email_resent_event(
     *, order: Order, user: UserType, email_type: OrderEventsEmails
 ) -> OrderEvent:
@@ -140,8 +184,25 @@ def order_manually_marked_as_paid_event(*, order: Order, user: UserType) -> Orde
     )
 
 
-def order_fully_paid_event(*, order: Order) -> OrderEvent:
-    return OrderEvent.objects.create(order=order, type=OrderEvents.ORDER_FULLY_PAID)
+def order_fully_paid_event(*, order: Order, user: UserType) -> OrderEvent:
+    if not _user_is_valid(user):
+        user = None
+    return OrderEvent.objects.create(
+        order=order, type=OrderEvents.ORDER_FULLY_PAID, user=user
+    )
+
+
+def payment_authorized_event(
+    *, order: Order, user: UserType, amount: Decimal, payment: Payment
+) -> OrderEvent:
+    if not _user_is_valid(user):
+        user = None
+    return OrderEvent.objects.create(
+        order=order,
+        type=OrderEvents.PAYMENT_AUTHORIZED,
+        user=user,
+        **_get_payment_data(amount, payment),
+    )
 
 
 def payment_captured_event(