+++ saleor/order/models.py
@@ -1,5 +1,6 @@
 from decimal import Decimal
 from operator import attrgetter
+from re import match
 from typing import Optional
 from uuid import uuid4
 
@@ -243,6 +244,15 @@ class Order(ModelWithMetadata):
             .exists()
         )
 
+    def is_captured(self):
+        return (
+            self.payments.filter(
+                is_active=True, transactions__kind=TransactionKind.CAPTURE
+            )
+            .filter(transactions__is_success=True)
+            .exists()
+        )
+
     @property
     def quantity_fulfilled(self):
         return sum([line.quantity_fulfilled for line in self])
@@ -458,6 +468,10 @@ class Fulfillment(ModelWithMetadata):
     def get_total_quantity(self):
         return sum([line.quantity for line in self])
 
+    @property
+    def is_tracking_number_url(self):
+        return bool(match(r"^[-\w]+://", self.tracking_number))
+
 
 class FulfillmentLine(models.Model):
     order_line = models.ForeignKey(
@@ -507,10 +521,3 @@ class OrderEvent(models.Model):
 
     def __repr__(self):
         return f"{self.__class__.__name__}(type={self.type!r}, user={self.user!r})"
-
-
-class Invoice(models.Model):
-    order = models.ForeignKey(Order, null=True, on_delete=models.SET_NULL)
-    number = models.CharField(max_length=255)
-    created = models.DateTimeField(auto_now_add=True)
-    url = models.URLField(max_length=2048)