+++ saleor/payment/interface.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from decimal import Decimal
-from typing import Any, Dict, Optional
+from typing import Any, Dict, List, Optional
 
 
 @dataclass
@@ -86,6 +86,7 @@ class GatewayConfig:
 
     gateway_name: str
     auto_capture: bool
+    supported_currencies: str
     # Each gateway has different connection data so we are not able to create
     # a unified structure
     connection_params: Dict[str, Any]
@@ -100,3 +101,13 @@ class CustomerSource:
     id: str
     gateway: str
     credit_card_info: Optional[CreditCardInfo] = None
+
+
+@dataclass
+class PaymentGateway:
+    """Dataclass for storing information about a payment gateway."""
+
+    id: str
+    name: str
+    currencies: List[str]
+    config: List[Dict[str, Any]]